SkyTrials Recode
Bump version to 2.0.0.0.
Drop <1.21.4 support.
Update the paper api to 1.21.7.
Update SkyLib to 1.3.0.0.
Drop the use ofpPaperweight in favor of the API now that there is a vault API.
Restructure the project to extend an AbstractTrial to support the creation of future trials.
Update the legacy trial to be a ChamberTrial.
Add a new LevelTrial.
Update player data to use a database rather than .yml files.
Add an EntityBuilder and configuration to support the creation of entities.
Added more configuration options to trials and trial spawners.
Update README FAQ and add information about future plans.
Update README description and features.
Updated license headers with an updated description.